微分

+++ forward_difference/forward_diff.c
@@ -13,7 +13,7 @@ double forward_diff(double h,double x) {
 
 int main(void) {
   double h,x;
-  double n;
+  int    n;
 
   //任意の値
   x = 1.0;
@@ -23,6 +23,7 @@ int main(void) {
 
   //f(x)の最上位bit
   h = pow( 2.0 , n );
+  //h = 2^n,2^(n-1),...2^(n-52)
   for (int i = 0; i < 53; i++) {
     //f'(x)を出力
     printf("[%2d]%.15lf\th=%.15e\n",i+1,forward_diff(h,x),h);
